Add files via upload
Handwritten digit recognition is a classic machine learning problem that has been studied for many years. The MNIST dataset is a popular dataset used for this task, and it contains 60,000 training images and 10,000 testing images of handwritten digits.

To build a handwritten digit recognition neural network in Python, you can use the Keras library. Keras is a high-level API for building neural networks, and it makes it easy to get started with deep learning.

Here are the steps involved in building a handwritten digit recognition neural network in Python using Keras:

Import the necessary libraries. Load the MNIST dataset. Preprocess the data. Create the model. Train the model. Evaluate the model. Make predictions.

This code will create a simple neural network with one hidden layer and 10 output neurons. The model will be trained for 10 epochs, and then it will be evaluated on the test set. The test accuracy of the model should be around 98%.

You can run this code in a Python environment. If you don't have a Python environment, you can install Anaconda, which is a distribution of Python that includes all the necessary libraries for machine learning